# docs/swagger.yaml
basePath: /
definitions:
  dtorequest.BarCreateRequest:
    properties:
      bar:
        type: string
      code:
        minLength: 3
        type: string
    required:
    - bar
    - code
    type: object
  dtorequest.BarUpdateRequest:
    properties:
      bar:
        type: string
      code:
        minLength: 3
        type: string
    required:
    - bar
    - code
    type: object
  dtorequest.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      rememberMe:
        type: boolean
    required:
    - email
    - password
    type: object
  dtorequest.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  dtoresponse.BarResponse:
    properties:
      bar:
        type: string
      code:
        type: string
      id:
        type: string
    type: object
  dtoresponse.LoginResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
      session:
        $ref: '#/definitions/dtoresponse.SessionResponse'
      tokens:
        $ref: '#/definitions/dtoresponse.TokenPairResponse'
      user:
        $ref: '#/definitions/dtoresponse.UserResponse'
    type: object
  dtoresponse.SessionResponse:
    properties:
      deviceId:
        type: string
      deviceName:
        type: string
      deviceType:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      ipAddress:
        type: string
      isActive:
        type: boolean
      lastUsedAt:
        type: string
      location:
        type: string
      userAgent:
        type: string
    type: object
  dtoresponse.TokenPairResponse:
    properties:
      accessToken:
        type: string
      accessTokenExpiresAt:
        type: string
      accessTokenExpiresIn:
        type: integer
      accessTokenType:
        type: string
      refreshToken:
        type: string
      refreshTokenExpiresAt:
        type: string
      refreshTokenExpiresIn:
        type: integer
      refreshTokenType:
        type: string
    type: object
  dtoresponse.UserResponse:
    properties:
      avatar:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      id:
        type: string
      isActive:
        type: boolean
      lastLoginAt:
        type: string
      name:
        type: string
    type: object
  response.BaseResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      pagination: {}
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  description: Go backend boilerplate using Clean Architecture. Provides a ready-to-use
    foundation for REST APIs with auth, RBAC, file uploads, and multi-database support.
  title: Goilerplate API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtorequest.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtoresponse.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /api/v1/auth/logout-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: Logout from all devices
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtoresponse.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtorequest.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Register a new user
      tags:
      - auth
  /api/v1/bars:
    get:
      parameters:
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtoresponse.BarResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: List bars
      tags:
      - bars
    post:
      consumes:
      - application/json
      parameters:
      - description: Bar data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtorequest.BarCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: Create bar
      tags:
      - bars
  /api/v1/bars/{id}:
    delete:
      parameters:
      - description: Bar ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: Delete bar
      tags:
      - bars
    get:
      parameters:
      - description: Bar ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtoresponse.BarResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get bar by ID
      tags:
      - bars
    put:
      consumes:
      - application/json
      parameters:
      - description: Bar ID
        in: path
        name: id
        required: true
        type: string
      - description: Bar data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtorequest.BarUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: Update bar
      tags:
      - bars
securityDefinitions:
  ApiKeyAuth:
    description: Partner API key.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"